+++ CHANGELOG.md
@@ -6,9 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [5.0.3] - 2023-04-30
-
+## Added
+-   ConnectionStringBuilder now accepts certificate path
+-   ConnectionStringBuilder now accepts a TokenCredential
 ### Fixed
-
 -   Fixed parsing of certain commands
 
 ## [5.0.2] - 2023-04-18